# backend/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Stages",
        "description": "Documentation de l'API Laravel pour gérer les étudiants, entreprises, tuteurs et stages.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local Laravel"
        }
    ],
    "paths": {
        "/api/entreprises": {
            "get": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Lister toutes les entreprises",
                "operationId": "712e9c332f3e4993d72988da80bea620",
                "responses": {
                    "200": {
                        "description": "Liste des entreprises récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Créer une entreprise",
                "operationId": "b9931edbf7734e3005034bc1f8c4d938",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "ville"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "BigBen Connected"
                                    },
                                    "secteur": {
                                        "type": "string",
                                        "example": "Technologie"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 Rue de Paris"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Fontainebleau"
                                    },
                                    "site_web": {
                                        "type": "string",
                                        "example": "https://www.bigben.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entreprise créée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/entreprises/{id}": {
            "get": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Récupérer une entreprise par ID",
                "operationId": "750bb8cbb174ff0a9ca9b20324af5a45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'entreprise",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entreprise trouvée"
                    },
                    "404": {
                        "description": "Entreprise non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Mettre à jour une entreprise",
                "operationId": "0efcf6f42950723824f2078295a4a839",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'entreprise",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Nouvelle Entreprise"
                                    },
                                    "secteur": {
                                        "type": "string",
                                        "example": "Informatique"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "456 Avenue de Lyon"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "site_web": {
                                        "type": "string",
                                        "example": "https://www.example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entreprise mise à jour avec succès"
                    },
                    "404": {
                        "description": "Entreprise non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Supprimer une entreprise",
                "operationId": "b8cde643f92887037e22ba472f77b539",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'entreprise",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Entreprise supprimée avec succès"
                    },
                    "404": {
                        "description": "Entreprise non trouvée"
                    }
                }
            }
        },
...
        "/api/tuteurs": {
            "post": {
                "tags": [
                    "Tuteurs"
                ],
                "summary": "Créer un tuteur",
                "operationId": "499de773ff760b8f91a3635997dce025",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "email",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Martin"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Pierre"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pierre.martin@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+33 6 12 34 56 78"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "ecole",
                                            "entreprise"
                                        ],
                                        "example": "entreprise"
                                    },
                                    "entreprise_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tuteur créé avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/tuteurs/{id}": {
            "put": {
                "tags": [
                    "Tuteurs"
                ],
                "summary": "Mettre à jour un tuteur",
                "operationId": "82a6335db3d21d48294a2531b3878b17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Martin"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Pierre"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pierre.martin@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tuteur mis à jour"
                    },
                    "404": {
                        "description": "Tuteur non trouvé"
                    }
                }
            }
        }
    },
...
}